fix(api-key): address review findings — security, types, i18n, tests, structure
Resolve all 13 actionable findings from code review on PR #341:

Blockers:
- Add missing RLS policy for api_keys table (AC-20)
- Add ParseUUIDPipe on revoke :id param
- Replace plain Error with BadRequestException in requireOrgId()
- Add resource:action regex validation on scopes Zod schema

Warnings:
- Remove `as never` casts on hasPermission (type already valid)
- Remove `as string` casts — route orgId through requireOrgId()
- Replace unsafe `as` cast for session permissions with type narrowing
- Replace hardcoded "Never" with i18n key (api_keys_never)
- Add orgId dependency to useApiKeys hook for org-switch re-fetch
- Add deriveStatus revoked+expired priority test
- Add empty scopes creation edge case test
- Add Zod validation schema edge case tests
- Split 766-line api-keys.tsx into modular structure

Refs #318
