+++ apps/web/src/routes/settings/api-keys/-types.ts
@@ -1,0 +1,1 @@
+export type ApiKeyStatus = 'active' | 'expired' | 'revoked'
